Fix calling sidekiq callbacks on failure and running without log file
When the `yield` block in `SidekiqRunner.run` fails and a `god_config`
callback was set the `cb` leaked into the `ensure` block. Since we want
to call a different callback there we have to use a different variable
name to not call `to_sym` on the proc instance `cb`.

When the `god_config` does not include a `log_file` (or it's explicitly
set to `nil`) we must not try to create the log files directory.